Migration to new separation model.
When you want to create a new client project, create a new repo and
clone it to wherever you want. All app assets go in an /assets folder
in that directory.

Run this command from <yourProjectRepoFolder>:

git submodule add https://github.com/ApigeeIgnite/ignite-iOS-engine
_ignite-iOS-engine

The resulting folder structure should then look like this:

<yourProjectRepoFolder>/_ignite-iOS-engine
<yourProjectRepoFolder>/assets

Caveat being that all changes to the engine still need to be made on
the master, since we shouldn’t be (and possibly can’t?) check things in
and out of a submodule.