See commit details.
Most of this commit is updates to use constant strings throughout the app.  Still need to work on this more but getting there.

Removed IXPropertyParser moving the parsing to the IXProperty and IXBaseShortCode classes.

Updated the IXBaseControl background.image code to improve it a little.  Also fixed a potential bug in the pinch handling code.

Fixed a bug where data providers would load before the controls in a view had their attributes applied and laid out.
Updated IXControlCacheContainer to add some functionality and work better.  The added functionality was for the Create action.