Added app data providers. Added the "clear_cache" function to the base data provider. Added a new concept of app level events.  See commit details for info.
Added "app_entered_foreground" and "app_entered_background" as application level events that can be listened to.  This is a little untested at the moment but works to an extent.  Note that the "push_recieved" app level event is not implemented yet but is listed in the constants file as one of them.  It will not work.