Added the "background_controls" node to the TableView control.  See details for further explanation.
Just like the "controls" node you now have the ability to add the node called "background_controls" to the TableView.

This allows for a background view to be added behind the cells main content.  In order for the background view to become visible you will have to add the "background_swipe_width" to the TableView control which specifies how far you want the cells top layer to slide over when revealing the background controls.