Updated the TableView and Collection controls.
The TableView and Collection controls now are subclasses of the IXCellBasedControl.  This allows us to share a lot of code that ended up being the same for both of these controls and stops us from having to write code twice.

The background_controls behavior has moved to a new class as well called the IXCellBackgroundSwipeController.  Basically the same reason for doing this as both the Collection and TableView cells will now be able to use the IXCellBackgroundSwipeController.