Changes include adding drawer controllers and new class to handle short code methods.
Added MMDrawerController pod while removing the JASidePanel pod.
Added the app level attributes "left_drawer_view" and "right_drawer_view" which specify the locations of the views that you want for the drawers.
Added the app functions "toggleDrawer.left" and "toggleDrawer.right" to be able to toggle the state of the drawers.
Added IXShortCodeFunction class that handles the IXBaseShortcode functions now instead of the mess that was in the IXBaseShortcode class.