Updated IXAppManager and IXImage.
Updated IXAppManager to use const strings for the attributes.
Updated IXAppManager to use IXPathHandler to get the local paths it needs.
Updated IXPathHandler to be able to grab local paths given a relative path with the files name.
Removed images.touch from the Image control.  Use the Button control to get these effects.