See Details.
Updated the TableView and Collection controls background controls to respond to touch when "background_slides_in_from_side" is turned on.

Added the "reset_all_background_controls" function to the TableView and Collection controls to force all the visible cells backgrounds to reset their positions.

Added "images.default.force_refresh" to the Image control so that you can tell it to refresh the cached image if its set to YES.

Updated the ActivityIndicator control so that it is centered on the controls content view.