Switched the layout engine back to the ObjC version. Fixed Button controls alpha attribute.  Fixed IXSize issue.
This is still pretty wonky.  Might need to rethink and reimplement IXSize and the alpha attribute.

Note also that the reason for changing the layout engine back from swift was not because it wasnt working properly but because of issues with cocoapod integration and swift. Once those are resolved we should/will go back to the Swift version.